refactor: change test folder (integration, unit -> domain)
